feat(player): don't collide with recruited beople (#13)
The problem was that each Berson prefab had two Capsule Colliders on it -- one explicit one attached to the GFX child and one implicit one that comes with the Character Controller attached to the root GameObject. I deleted the child Capsule Collider and reconfigured the one on the Character Controller to do what we want.